Switch from printk to cprintf

+++ src/kern/io/console.c
@@ -43,7 +43,7 @@ void
 cprintf(char *fmt, ...)
 {
   int i, c, locking;
-  uint *argp;
+  ulong *argp;
   char *s;
 
   locking = cons.locking;
@@ -53,7 +53,7 @@ cprintf(char *fmt, ...)
   if (fmt == 0)
     panic("null fmt");
 
-  argp = (uint*)(void*)(&fmt + 1);
+  argp = (ulong*)(void*)(&fmt + 1);
   for(i = 0; (c = fmt[i] & 0xff) != 0; i++){
     if(c != '%'){
       consputc(c);
